+++ ntex/src/lib.rs
@@ -84,8 +84,6 @@ pub mod server {
     //! General purpose tcp server
     pub use ntex_server::net::*;
 
-    pub use ntex_server::{Signal, signal};
-
     #[cfg(feature = "openssl")]
     pub use ntex_tls::openssl;
 
@@ -124,10 +122,6 @@ pub mod util {
     pub use ntex_bytes::{Buf, BufMut, ByteString, Bytes, BytesMut};
     pub use ntex_bytes::{BytePage, BytePageSize, BytePages};
     pub use ntex_util::{HashMap, HashSet, future::*, services::*};
-
-    #[doc(hidden)]
-    #[allow(deprecated)]
-    pub use ntex_util::error::*;
 
     #[doc(hidden)]
     pub fn enable_test_logging() {