+++ ntex-dispatcher/CHANGES.md
@@ -1,5 +1,9 @@
 # Changes
 
+## [4.0.0] - 2026-08-xx
+
+* Migrate to ntex-service 5
+
 ## [3.2.1] - 2026-06-15
 
 * Send `Control::WBackPressureEnabled` control message only once per back-pressure